Velbus split of entity in its own file
