+++ homeassistant/components/velbus/__init__.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 
 import logging
 
-from velbusaio.channels import Channel as VelbusChannel
 from velbusaio.controller import Velbus
 import voluptuous as vol
 
@@ -13,7 +12,6 @@ from homeassistant.core import HomeAssistant, ServiceCall
 from homeassistant.helpers import device_registry
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.device_registry import DeviceEntry
-from homeassistant.helpers.entity import DeviceInfo, Entity
 
 from .const import (
     CONF_INTERFACE,
@@ -146,32 +144,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.services.async_remove(DOMAIN, SERVICE_SYNC)
         hass.services.async_remove(DOMAIN, SERVICE_SET_MEMO_TEXT)
     return unload_ok
-
-
-class VelbusEntity(Entity):
-    """Representation of a Velbus entity."""
-
-    _attr_should_poll: bool = False
-
-    def __init__(self, channel: VelbusChannel) -> None:
-        """Initialize a Velbus entity."""
-        self._channel = channel
-        self._attr_name = channel.get_name()
-        self._attr_device_info = DeviceInfo(
-            identifiers={
-                (DOMAIN, str(channel.get_module_address())),
-            },
-            manufacturer="Velleman",
-            model=channel.get_module_type_name(),
-            name=channel.get_full_name(),
-            sw_version=channel.get_module_sw_version(),
-        )
-        serial = channel.get_module_serial() or str(channel.get_module_address())
-        self._attr_unique_id = f"{serial}-{channel.get_channel_number()}"
-
-    async def async_added_to_hass(self) -> None:
-        """Add listener for state changes."""
-        self._channel.on_status_update(self._on_update)
-
-    async def _on_update(self) -> None:
-        self.async_write_ha_state()
